Add function to common, extract paths into own module and make changes so it still works. Add some fixmes/todos.
