+++ common.js
@@ -3,20 +3,19 @@ var sinon = require("sinon");
 var sinonChai = require("sinon-chai");
 chai.should();
 chai.use(sinonChai);
-
-
 exports.chai = chai;
 exports.sinon = sinon;
-
 exports.assert = chai.assert;
 exports.expect = chai.expect;
 
 exports.gameDir = '../game/';
-
 exports.levels = ['mine']; //TODO: put this in maps module
 
 exports.isUndefined = isUndefined;
 exports.extend = extend;
+exports.noop = noop;
+
+function noop(){}
 
 function extend(obj, extension) {
     obj = obj || {};
