+++ game/combat.js
@@ -1,6 +1,8 @@
 var RNG = require('rot-js').RNG;
 
 //TODO: Add potential to miss based on attacker's sight/speed and defender's speed/defense.
+//FIXME: -1 damage can be dealt. Damage should never be lower than 0.
+
 
 /*
  *  Returns a message object with the results of combat.
